Divided the class TelegramBotClient into parts
