+++ NetTelebot/TelegramBotClient.cs
@@ -1,9 +1,6 @@
 ﻿using RestSharp;
 using System;
-using System.Linq;
-using System.Net;
 using System.Runtime.CompilerServices;
-using System.Threading;
 using NetTelebot.BotEnum;
 using NetTelebot.Interface;
 using NetTelebot.Result;
@@ -32,35 +29,14 @@ namespace NetTelebot
      * There are requests to the telegram servers
      */
 
-    /// <summary>
-    /// The main class to use Telegram Bot API. Get an instance of this class and set the Token property and start calling methods.
-    /// </summary>
-    public class TelegramBotClient
+    /* About this partial class
+     * 
+     * in this part of the class, only the methods that work with the api telegram.
+     * 
+     */
+
+    public partial class TelegramBotClient
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="TelegramBotClient"/> class.
-        /// </summary>
-        public TelegramBotClient()
-        {
-            CheckInterval = 1000;
-            RestClient = new RestClient("https://api.telegram.org");
-        }
-
-        /// <summary>
-        /// Your bot token
-        /// </summary>
-        public string Token { get; set; }
-
-        /// <summary>
-        /// Gets or sets the REST client. Used in integartion test.
-        /// </summary>
-        internal RestClient RestClient { private get; set; }
-
-        /// <summary>
-        /// Interval time in milliseconds to get latest messages sent to your bot.
-        /// </summary>
-        public int CheckInterval { get; set; }
-
         private const string getMeUri = "/bot{0}/getMe";
         private const string getUpdatesUri = "/bot{0}/getUpdates";
         private const string sendMessageUri = "/bot{0}/sendMessage";
@@ -88,88 +64,11 @@ namespace NetTelebot
         private const string answerCallbackQueryUri = "/bot{0}/answerCallbackQuery";
         private const string answerShippingQueryUri = "/bot{0}/answerShippingQuery";
 
-        private Timer mUpdateTimer;
-        private int mLastUpdateId;
-
-        /// <summary>
-        /// Occurs when [get updates error].
-        /// </summary>
-        public event UnhandledExceptionEventHandler GetUpdatesError;
-
-        /// <summary>
-        /// Whenever a message is sent to your bot, this event will be raised.
-        /// </summary>
-        public event EventHandler<TelegramUpdateEventArgs> UpdatesReceived;
-
-        /// <summary>
-        /// Gets first 100 messages sent to your bot.
-        /// </summary>
-        /// <returns>Returns a class containing messages sent to your bot</returns>
-        public GetUpdatesResult GetUpdates()
-        {
-            return GetUpdatesInternal(null, null);
-        }
-
-        /// <summary>
-        /// Gets maximum 100 messages sent to your bot, starting from update_id set by offset
-        /// </summary>
-        /// <param name="offset">First update_id to be downloaded</param>
-        /// <returns>On success, the sent <see cref="GetUpdatesResult"/> is returned.</returns>
-        public GetUpdatesResult GetUpdates(int offset)
-        {
-            return GetUpdatesInternal(offset, null);
-        }
-
-        /// <summary>
-        /// Gets messages sent to your bot, starting from update_id set by offset, maximum number is set by limit
-        /// </summary>
-        /// <param name="offset">First update_id to be downloaded</param>
-        /// <param name="limit">Maximum number of messages to receive. It cannot be more than 100</param>
-        /// <returns>On success, the sent <see cref="GetUpdatesResult"/> is returned.</returns>
-        public GetUpdatesResult GetUpdates(int offset, byte limit)
-        {
-            return GetUpdatesInternal(offset, limit);
-        }
-
-        /// <summary>
-        /// Gets messages sent to your bot, from the begining and maximum number of limit set as parameter
-        /// </summary>
-        /// <param name="limit">Maximum number of messages to receive. It cannot be more than 100</param>
-        /// <returns>Returns a class containing messages sent to your bot</returns>
-        public GetUpdatesResult GetUpdates(byte limit)
-        {
-            return GetUpdatesInternal(null, limit);
-        }
-
-        //todo refact this
-        private GetUpdatesResult GetUpdatesInternal(int? offset, byte? limit)
-        {
-            CheckToken();
-
-            RestRequest request = new RestRequest(string.Format(getUpdatesUri, Token), Method.POST);
-
-            if (offset.HasValue)
-                request.AddQueryParameter("offset", offset.Value.ToString());
-            if (limit.HasValue)
-                request.AddQueryParameter("limit", limit.Value.ToString());
-
-            return ExecuteRequest<GetUpdatesResult>(request) as GetUpdatesResult;
-        }
-
         private RestRequest NewRestRequest(string uri)
         {
             RestRequest request = new RestRequest(string.Format(uri, Token), Method.POST);
 
             return request;
-        }
-
-        /// <summary>
-        /// Called when [get updates error].
-        /// </summary>
-        /// <param name="exception">The exception.</param>
-        protected virtual void OnGetUpdatesError(Exception exception)
-        {
-            GetUpdatesError?.Invoke(this, new UnhandledExceptionEventArgs(exception, false));
         }
 
         /// <summary>
@@ -858,38 +757,6 @@ namespace NetTelebot
 
         //todo Inline mode methods (https://core.telegram.org/bots/api#inline-mode-methods)
 
-        private void CheckToken()
-        {
-            if (Token == null)
-                throw new Exception("Token is null");
-        }
-
-        /// <summary>
-        /// Checks new updates (sent messages to your bot) automatically. Set CheckInterval property and handle UpdatesReceived event.
-        /// </summary>
-        public void StartCheckingUpdates()
-        {
-            CheckToken();
-
-            if (mUpdateTimer == null)
-            { 
-                mUpdateTimer = new Timer(UpdateTimerCallback, null, CheckInterval, Timeout.Infinite);
-            }
-            else
-            {
-                mUpdateTimer.Change(CheckInterval, Timeout.Infinite);
-            }
-        }
-
-        /// <summary>
-        /// Stops automatic checking updates
-        /// </summary>
-        public void StopCheckUpdates()
-        {
-            mUpdateTimer?.Dispose();
-            mUpdateTimer = null;
-        }
-
         private static RestRequest AddFile(IFile iFile, RestRequest request, string name)
         {
             ExistingFile file = iFile as ExistingFile;
@@ -909,85 +776,6 @@ namespace NetTelebot
             }
 
             return request;
-        }
-
-        private void UpdateTimerCallback(object state)
-        {
-            GetUpdatesResult updates = null;
-            var getUpdatesSuccess = false;
-
-            try
-            {
-                updates = mLastUpdateId == 0
-                    ? GetUpdates()
-                    : GetUpdates(mLastUpdateId + 1);
-
-                getUpdatesSuccess = true;
-            }
-            catch (Exception ex)
-            {
-                OnGetUpdatesError(ex);
-            }
-
-            if (getUpdatesSuccess)
-
-                if (updates.Ok && updates.Result != null && updates.Result.Any())
-                {
-                    mLastUpdateId = updates.Result.Last().UpdateId;
-                    OnUpdatesReceived(updates.Result);
-                }
-
-            mUpdateTimer?.Change(CheckInterval, Timeout.Infinite);
-        }
-
-        /// <summary>
-        /// Called when [updates received].
-        /// </summary>
-        /// <param name="updates">The updates</param>
-        protected virtual void OnUpdatesReceived(UpdateInfo[] updates)
-        {
-            TelegramUpdateEventArgs args = new TelegramUpdateEventArgs(updates);
-            UpdatesReceived?.Invoke(this, args);
-        }
-
-        private object ExecuteRequest<T>(IRestRequest request) where T : class
-        {
-            IRestResponse response = RestClient.Execute(request);
-
-            if (response.StatusCode == HttpStatusCode.OK)
-            {
-                if (typeof(T) == typeof (SendMessageResult))
-                    return new SendMessageResult(response.Content);
-
-                if (typeof (T) == typeof (BooleanResult))
-                    return new BooleanResult(response.Content);
-
-                if (typeof(T) == typeof(UserInfoResult))
-                    return new UserInfoResult(response.Content);
-
-                if (typeof(T) == typeof (GetUserProfilePhotosResult))
-                    return new GetUserProfilePhotosResult(response.Content);
-
-                if (typeof(T) == typeof (GetUpdatesResult))
-                    return new GetUpdatesResult(response.Content);
-
-                if (typeof(T) == typeof (ChatInfoResult))
-                    return new ChatInfoResult(response.Content);
-
-                if (typeof(T) == typeof (IntegerResult))
-                    return new IntegerResult(response.Content);
-
-                if(typeof(T) == typeof(FileInfoResult))
-                    return new FileInfoResult(response.Content);
-
-                if(typeof(T) == typeof(ChatMembersInfoResult))
-                    return new ChatMembersInfoResult(response.Content);
-
-                if (typeof(T) == typeof(ChatMemberInfoResult))
-                    return new ChatMemberInfoResult(response.Content);
-            }
-
-            throw new Exception(response.StatusDescription);
         }
     }
 }
